+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 * Fix bug in `UrlQuotes` linter where lint would be reported for data URIs
 * Add `SelectorFormat` linter which checks that the names of ids, classes, etc.
   in selectors match a desired convention
+* Remove `CapitalizationInSelector`, which has been superseded by the more
+  powerful `SelectorFormat` linter
 
 ## 0.28.0